update UI manager management, update code with new db, add crud movies

+++ src/main/java/org/group3/project_swp391_bookingmovieticket/entities/Room.java
@@ -11,14 +11,14 @@ public class Room {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private int capacity;
+    private Integer capacity;
 
     private String imgurl;
 
     private String name;
 
     @Column(name = "total_area")
-    private double totalArea;
+    private Float totalArea;
 
     @ManyToOne
     @JoinColumn(name = "branch_id")
